refactor(tests): nest performance, platform, settings, transcode, updates service tests
