+++ tests/unit/renderer/infrastructure/services/settings/settings-cinematic-mode.service.test.ts
@@ -5,8 +5,8 @@
 import { describe, it, expect, beforeEach, vi } from 'vitest';
 import { SettingsCinematicModeService } from '@renderer/infrastructure/services/settings/settings-cinematic-mode.service';
 import { EventChannels } from '@platform/events';
-import { createAppState } from '../../../../factories/index.js';
-import { createInjectableHarness } from '../../../../support/di/injectable.harness.js';
+import { createAppState } from '../../../../../factories/index.js';
+import { createInjectableHarness } from '../../../../../support/di/injectable.harness.js';
 
 describe('SettingsCinematicModeService', () => {
   let service;
